Move webpack config for sass to a separate directory
Sass files will be compiled into CSS files into the same folder (/static/css/) as it was done with libsass and setuptools.
The resulting files in /static/css will have same layout and names as with the previous solution, except for /static/css/foundation folder. It is dropped completely since it was empty when built with setuptools. This is expected since /sass/foundation folder contains only Sass partials.